[wtrack] Implementing form step 2

+++ wtrack/forms.py
@@ -25,8 +25,8 @@ class RecordForm(forms.ModelForm):
 
     date = forms.DateField(widget=forms.SelectDateWidget(), required=True)
     ## day = models.CharField(max_length=3, default=day_of_week(date))
-    sleep_hours = forms.SmallIntegerField()
-    calories_consumed = forms.SmallIntegerField()
+    sleep_hours = forms.IntegerField()
+    calories_consumed = forms.IntegerField()
     climbing_flag = forms.CharField(max_length=2, widget=forms.Select())#choices=CLIMBING_FLAGS)
     gym_flag = forms.CharField(max_length=2, widget=forms.Select())#choices=GYM_FLAGS)
     alco_flag = forms.CharField(max_length=2, widget=forms.Select(), required=True)#choices=ALCO_FLAGS)
